refactor(video): split show_chip() function
Prior to this commit, the name of the `show_chip()` function was
misleading.  The `show_chip()` function actually was **NOT** responsible
for _showing_ the video frames; it was only responsible for the
following operations:

 - read video frames from the webcam
 - detect AruCo markers in each frame, and draw overlay to indicate
   markers (if available)
 - apply perspective correction based on detected AruCo marker positions
   (if applicable)
 - detect chip UUID from QR code (if available)
 - combine raw video frame and perspective-corrected frame into a single
   frame

The display of the video frames was _actually_ performed by the `main()`
function (which also wrote the UUID as text in top-left corner of the
combined video frame).

In this commit, refactor as follows:
 - Rename `show_chip()` function as `chip_video_process()`.
 - Rename `main()` function as `show_chip()` and do _not_ launch thread
   from within `show_chip()`.
 - Add new `main()` function to launch `chip_video_process()` in a
   thread and call `show_ship()`.