feat(test): add video-dir CLI argument
Prior to this commit, upon test completion, a search was conducted to
look for a recorded video that corresponded with the time of the start
of the test within the `~/Videos/Captures` directory.

In this commit, add the optional `--video-dir` to explicitly specify the
directory to search for videos; rather than using the hard-coded
`~/Videos/Captures` directory.