feat(test): skip unreachable electrodes
During the execution of a chip test, an electrode on the test route _may
become unreachable_.  This can happen, e.g., if a bottleneck electrode
has failed, which cutting off the only path to other electrodes on the
test route.

Prior to this commit, if an unreachable electrode was encountered during
a chip test, the test was terminated immediately.

In this commit, emit an event using the ``electrode-skip`` signal if an
unreachable electrode is encountered and continue with the remainder of
the test route.  Maintain a list of all skipped electrodes during the
test and include it in the HTML output report.