fix(connect): force downcast of mm x/y to float
Prior to this commit, execution of the `connect.load_device()` function
with `pint==0.9` and `pandas==0.22.0` resulted in the following error:

```
C:\Users\chris\mc2-x86\envs\db-chip-qc\lib\site-packages\pint\quantity.py:1377: UnitStrippedWarning: The unit of the quantity is stripped.
  warnings.warn("The unit of the quantity is stripped.", UnitStrippedWarning)
```

In this commit, force downcast of `pint` millimeter quantities for `x`
and `y` coordinates to `float` to work around the issue.