feat(test): accept test route id or waypoints list
Prior to this commit, the only way to explicitly specify a test route
to the `dropbot_chip_qc.bin.test` script was to provide a JSON list of
channel numbers, e.g., `"[110, 109, 115]"`.

In this commit, add support for reading embedded test route waypoints
from the chip design SVG file (see sci-bots/dmf-chip#1), allowing a test
route to be specified as either an id of a `<dmf:TestRoute>` element in
the SVG _or_ a JSON list of channel numbers.