feat(test): log chip info to `test-start` message
Prior to this commit, no information about the chip design was included
in the output report JSON log.  Only a rendered image of the chip layout
indicating the test route was included in the HTML report.

In this commit, add the chip information returned by `dmf_chip.load()`
to the log; which is JSON-serialized and included in the output report.