feat(render): draw chip layout using dmf_chip
Prior to this commit, the `dropbot.chip.draw()` was used to draw the
electrode layout in rendered test reports.  However, the `dropbot.chip`
sub-module has been deprecated in favour of the sci-bots/dmf-chip.

In this commit, use the `dmf_chip.draw()` function to draw the chip
electrode layout in rendered test reports.