feat(notebooks/ui): use MQTT DropBot proxy
Prior to this commit, the PySide2 quality control user interface used a
direct serial connection to the DropBot, through a DropBot monitor
process. However, this prevented interaction with the DropBot from other
apps, e.g., to debug situations where a drop is stuck.

In this commit, refactor the PySide2 quality control user interface to
use a DropBot MQTT proxy, which connects to an external DropBot monitor
system tray app. This decouples the DropBot connection from the PySide2
quality control UI.