refactor(ui): add ExecutorController
Prior to this commit, the code for managing the state of a
`dropbot_chip_qc.ui.execute.Executor` instance was contained within the
`dropbot_chip_qc.ui.notebook.executor_control()` function. However, this
limited control to within a Jupyter notebook due to a dependence on
IPython widgets.

In this commit, refactor the code from the `executor_control()` function
that does not depend on IPython widgets into a new `ExecutorController`
class within the `dropbot_chip_qc.ui.execute` module. This will make it
possible to share logic that is not tied to a specific UI framework.