refactor(ui/notebook): use ExecutorController
Prior to this commit, the `executor_control()` function in the
`dropbot_chip_qc.ui.notebook` module implemented all logic required
for managing the state of an `Executor` instance.

In this commit, refactor the `executor_control()` function to use the
new `ExecutorController` class for code that is not specific to IPython
widgets.